+++ packages/browser-wallet/src/popup/pages/Account/Account.tsx
@@ -19,7 +19,7 @@ import TransactionLog from './TransactionLog/TransactionLog';
 import SendCcd from './SendCcd';
 import ConnectedBox from './ConnectedBox';
 import Tokens from './Tokens';
-import AddTokens from './AddTokens';
+import ManageTokens from './ManageTokens';
 import { AccountPageContext, accountPageContext } from './utils';
 
 function Account() {
@@ -93,7 +93,7 @@ export default function AccountRoutes() {
                     <Route path={`${accountRoutes.log}/*`} element={<TransactionLog />} />
                     <Route path={`${accountRoutes.settings}/*`} element={<AccountSettings />} />
                     <Route path={`${accountRoutes.tokens}/*`} element={<Tokens />} />
-                    <Route path={`${accountRoutes.manageTokens}/*`} element={<AddTokens />} />
+                    <Route path={`${accountRoutes.manageTokens}/*`} element={<ManageTokens />} />
                 </Route>
             </Routes>
         </accountPageContext.Provider>
