--- packages/browser-wallet/src/popup/pages/Account/AddTokens/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AddTokens';
