+++ packages/browser-wallet/src/popup/pages/Account/ManageTokens/ManageTokens.scss
