+++ packages/browser-wallet/src/popup/pages/Account/ManageTokens/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ManageTokens';
