+++ packages/browser-wallet/src/popup/pages/Account/index.scss
@@ -6,4 +6,4 @@
 @import './TransactionLog/TransactionLog';
 @import './SendCcd/SendCcd';
 @import './Tokens/Tokens';
-@import './AddTokens/AddTokens';
+@import './ManageTokens/ManageTokens';
